+++ D2ApiViewerEditor/D2ViewerEditor.Api.UnitTests/Middleware/RequestObservabilityMiddlewareTests.cs
@@ -1,0 +1,41 @@
+using D2ViewerEditor.Api.Middleware;
+using FluentAssertions;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Logging.Abstractions;
+using NUnit.Framework;
+
+namespace D2ViewerEditor.Api.UnitTests.Middleware;
+
+[TestFixture]
+public class RequestObservabilityMiddlewareTests
+{
+    private static RequestObservabilityMiddleware Build(RequestDelegate next) =>
+        new(next, NullLogger<RequestObservabilityMiddleware>.Instance);
+
+    [Test]
+    public async Task Sets_correlation_id_response_header_when_absent()
+    {
+        var context = new DefaultHttpContext();
+        var middleware = Build(_ => Task.CompletedTask);
+
+        await middleware.InvokeAsync(context);
+
+        context.Response.Headers[RequestObservabilityMiddleware.CorrelationIdHeader].ToString()
+            .Should().NotBeNullOrWhiteSpace();
+        context.Items[RequestObservabilityMiddleware.CorrelationIdItemKey].Should().NotBeNull();
+    }
+
+    [Test]
+    public async Task Echoes_incoming_correlation_id()
+    {
+        var context = new DefaultHttpContext();
+        context.Request.Headers[RequestObservabilityMiddleware.CorrelationIdHeader] = "corr-abc";
+        var middleware = Build(_ => Task.CompletedTask);
+
+        await middleware.InvokeAsync(context);
+
+        context.Response.Headers[RequestObservabilityMiddleware.CorrelationIdHeader].ToString()
+            .Should().Be("corr-abc");
+        context.Items[RequestObservabilityMiddleware.CorrelationIdItemKey].Should().Be("corr-abc");
+    }
+}