+++ D2ApiViewerEditor/D2ViewerEditor.Api/Extensions/MiddlewareExtensions.cs
@@ -8,4 +8,9 @@ public static class MiddlewareExtensions
     {
         return app.UseMiddleware<ExceptionHandlingMiddleware>();
     }
+
+    public static IApplicationBuilder UseRequestObservability(this IApplicationBuilder app)
+    {
+        return app.UseMiddleware<RequestObservabilityMiddleware>();
+    }
 }